chore: :adhesive_bandage: Fix running scripts with yarn
This fixes a bug when running scripts with yarn.
The error message is "error TypeError [ERR_INVALID_ARG_VALUE]:
The property 'options.env['npm_package_description']' must be a string without null bytes."

This is due to the fact that for some reason, when the "description"
field is missing in the package.json, by defaults it loads the first line
of the README. Which in our case, is a binary file :-( (README.jpg).

By specifying a description in the package.json, we avoid this issue.