feat: :sparkles: Level 1 - Static table
This is simply the level 1 static table, using the data file in the `data` folder.

I also introduced CSS Modules to the existing app, in order to bring some
structure and isolation to CSS styles.

Finally, I separated it into a pseudo-page, even though we do not have a router now.

fix: :rotating_light: Fix the build errors with CSS Modules

There were some errors with CSS syntax and tsc with CSS Modules when creating the final build.
This fixes those minor issues.