feat: :hammer: Add mock API
This adds a very simple API integrated with vite dev server.
It serves the files in the `data` folder in the subpath `/api`.
Being in the subpath we do not have to worry about CORS.

This is implemented using a dodgy package that I would not recommend in a "real" setting.

To test it:
`curl -i http://localhost:5173/api/monthlyTargets.json`