feat: :sparkles: Level 3 - add editable cells + re-computing of targets
I chose to use a simple reducer approach to manage the state of the app.
All computing has been moved to the `organization-targets` module.
I also added a few tests to make sure the app is working as expected.

BTW - I noticed that the sample data is not consistent with the requirements.
This means that either my understanding of the requirements is wrong or the sample data is wrong.
We'll have to figure out with stakeholders.

Also - this uses a contenteditable td - which is not the best and may need an improvement in the future.