feat: :sparkles: Level 4 - Add quater columns
This simply builds on top of the previous work and adds the quater columns to the grid.
No fundamental architecture changes were needed, just added the relevant type declaration and computation.

One thing to note: I did not apply the spec exactly as is felt weird.
The spec says that the AverageBeginningMRR should be inverted (`* -1`), but:
 - I don't understand why
 - It does not fit with the sample data
This will need to be confirmed with stakeholders. I left a TODO.

I also fixed some styling issues with light/dark mode.